Add BRCK theming
The navigation bar has been modified so that is aligns
with the BRCK look and feel. The ThingSpeak logo has
also been replaced with the BRCK logo

I have also stripped down the home page so that only the
stock banner image and a login button are displayed.
Additionally, I have modified the layout of the sign up
and registration pages so that the forms are centered on
the page

Signed-off-by: Emmanuel Kala <emmanuel@brck.com>